# docs/architecture/api-specification.yaml
openapi: 3.0.3
info:
  title: FitrahTube API
  version: 1.0.0
  description: |
    REST API for FitrahTube backend. Production implementation with 113+ endpoints across 14 controllers.

    **Key Features:**
    - Public content API for Android app (no auth required)
    - Admin/moderator content management with approval workflows
    - YouTube integration for content discovery
    - Bulk import/export functionality
    - Video validation and health monitoring
    - Comprehensive audit logging

    **Authentication:**
    - Firebase Authentication with JWT tokens
    - Role-based access: ADMIN, MODERATOR
    - Public endpoints require no authentication

    **Pagination:**
    - Cursor-based pagination for infinite scroll
    - Page-based pagination for admin interfaces

    **Caching:**
    - Categories: 1 hour TTL
    - Channels: 15 minutes TTL
    - YouTube searches: Varies by type
servers:
  - url: http://localhost:8080/api
    description: Local Development
  - url: http://192.168.1.167:8080/api
    description: Local Network (Android Testing)
  - url: https://api.albunyaan.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /admin/sort/categories/reorder:
    put:
      summary: Reorder a category (insert-and-shift)
      operationId: reorderCategory
      tags: [Sort Order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, newPosition]
              properties:
                categoryId:
                  type: string
                newPosition:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Updated category sort order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategorySortDto'
  /admin/sort/categories/{categoryId}/content:
    get:
      summary: Get content items in sort order within a category
      operationId: getCategoryContentOrder
      tags: [Sort Order]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content items sorted by position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentSortDto'
  /admin/sort/categories/{categoryId}/content/reorder:
    put:
      summary: Reorder content within a category (insert-and-shift)
      operationId: reorderContentInCategory
      tags: [Sort Order]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentId, contentType, newPosition]
              properties:
                contentId:
                  type: string
                contentType:
                  type: string
                  enum: [channel, playlist, video]
                newPosition:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Updated content sort order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentSortDto'

  # ============================================================================
  # CATEGORIES (Admin/Moderator)
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication JWT token
  schemas:
...
    CategorySortDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        localizedNames:
          type: object
          additionalProperties:
            type: string
          nullable: true
        displayOrder:
          type: integer
          nullable: true
        contentCount:
          type: integer
    ContentSortDto:
      type: object
      properties:
        contentId:
          type: string
        contentType:
          type: string
          enum: [channel, playlist, video]
        title:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        position:
          type: integer
        youtubeId:
          type: string
...